web/backend/requirements.txt
pymongo==4.6.1
python-dotenv==1.0.0
pydantic==2.6.1
click==8.1.7 